build mvn test script

+++ dev/mvn_test.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 
+#mlsql-build-env
 function exit_with_usage {
   cat << EOF
 usage: mvn_test
@@ -9,7 +10,6 @@ Inputs are specified with the following environment variables:
 GIT_BRANCH - Git branch on which to make release
 SCRIPT_FILE - the script file you want execute in remote server
 MLSQL_SPARK_VERSIOIN - the spark version
-WITHOUT_STOP - whether to stop the instance
 
 EOF
   exit 1
@@ -22,29 +22,88 @@ if [[ $@ == *"help"* ]]; then
   exit_with_usage
 fi
 
-for env in GIT_BRANCH SCRIPT_FILE MLSQL_SPARK_VERSIOIN WITHOUT_STOP; do
+for env in GIT_BRANCH SCRIPT_FILE MLSQL_SPARK_VERSIOIN; do
   if [ -z "${!env}" ]; then
     echo "$env must be set to run this script; \n Please run ./mvn_test.sh help to get how to use."
     exit 1
   fi
 done
 
+
+if [ -z "$1" ]; then
+    echo "$env must be set to run this script; \n Please run ./mvn_test.sh help to get how to use."
+    exit 1
+fi
+
+echo "install git in remote server"
+cat << EOF > ${SCRIPT_FILE}
+#!/usr/bin/env bash
+apt-get install git -y -q
+cd /home/webuser
+rm streamingpro.tar.gz
+rm -rf streamingpro
+rm temp_ServiceFramework.tar.gz
+rm -rf temp_ServiceFramework
+EOF
+
+python -m pymlsql.aliyun.dev.run_remote_shell --script_path ${SCRIPT_FILE} --keyPairName "mlsql-build-env-local" --execute_user root --instance_id $1
+
+echo "download  ServiceFramework and copy to remote"
+rm -rf temp_ServiceFramework
+git clone --depth 1 https://github.com/allwefantasy/ServiceFramework.git temp_ServiceFramework
+tar czvf temp_ServiceFramework.tar.gz temp_ServiceFramework
+python -m pymlsql.aliyun.dev.copy_from_local --source temp_ServiceFramework.tar.gz --target /home/webuser --keyPairName "mlsql-build-env-local" --execute_user root --instance_id $1
+
+
+
+MLSQL_HTTPS_REPO="https://github.com/allwefantasy/streamingpro.git"
+MLSQL_GIT_REPO="git@github.com:allwefantasy/streamingpro.git"
+
+echo "download streamingpro and copy to remote server"
+rm -rf streamingpro
+git clone ${MLSQL_GIT_REPO} -b ${GIT_BRANCH}
+tar czvf streamingpro.tar.gz streamingpro
+
+python -m pymlsql.aliyun.dev.copy_from_local --source streamingpro.tar.gz --target /home/webuser --keyPairName "mlsql-build-env-local" --execute_user root --instance_id $1
+
+
+echo "grant the resource to webuser"
+cat << EOF > ${SCRIPT_FILE}
+#!/usr/bin/env bash
+
+cd /home/webuser
+chown -R webuser:webuser streamingpro.tar.gz
+chown -R webuser:webuser temp_ServiceFramework.tar.gz
+
+EOF
+
+python -m pymlsql.aliyun.dev.run_remote_shell --script_path ${SCRIPT_FILE} --keyPairName "mlsql-build-env-local" --execute_user root --instance_id $1
+
+echo "unzip the resource"
+cat << EOF > ${SCRIPT_FILE}
+#!/usr/bin/env bash
+
+source activate mlsql-3.5
+cd /home/webuser
+tar xvf temp_ServiceFramework.tar.gz
+tar xvf streamingpro.tar.gz
+EOF
+
+python -m pymlsql.aliyun.dev.run_remote_shell --script_path ${SCRIPT_FILE} --keyPairName "mlsql-build-env-local" --execute_user webuser --instance_id $1
+
+echo "mvn install ServiceFramework and test streamingpro"
+
 cat << EOF > ${SCRIPT_FILE}
 #!/usr/bin/env bash
 
 source activate mlsql-3.5
 
-rm -rf /tmp/temp_ServiceFramework
-git clone --depth 1 https://github.com/allwefantasy/ServiceFramework.git /tmp/temp_ServiceFramework
-cd /tmp/temp_ServiceFramework
-mvn install -DskipTests -Pjetty-9 -Pweb-include-jetty-9
+cd /home/webuser/temp_ServiceFramework
 
-MLSQL_HTTPS_REPO="https://github.com/allwefantasy/streamingpro.git"
-MLSQL_GIT_REPO="git@github.com:allwefantasy/streamingpro.git"
+echo "install ServiceFramework"
+mvn install -DskipTests -Pjetty-9 -Pweb-include-jetty-9 > sf.log
 
-rm -rf streamingpro
-git clone \${MLSQL_GIT_REPO} -b ${GIT_BRANCH}
-cd streamingpro
+cd /home/webuser/streamingpro
 
 BASE_PROFILES="-Pscala-2.11 -Ponline -Phive-thrift-server -Pcarbondata  -Pcrawler -Pautoml"
 PUBLISH_SCALA_2_10=0
@@ -57,25 +116,12 @@ fi
 
 BASE_PROFILES="\$BASE_PROFILES -Pspark-$MLSQL_SPARK_VERSIOIN -Pstreamingpro-spark-$MLSQL_SPARK_VERSIOIN-adaptor"
 
-
-mvn test -pl streamingpro-mlsql -am \$BASE_PROFILES
-
-EOF
-
-
-
-
-
-
-# here we will create a ECS instance dynamically and run the script file
-# the remote server.
-cat << EOF
-execute python command:
-        python -m pymlsql.aliyun.dev.run_remote_shell --script_path ${SCRIPT_FILE}
+echo "test streamingpro"
+mvn test -pl streamingpro-mlsql -am \$BASE_PROFILES > sg-test.log
 
 EOF
 
 
-python -m pymlsql.aliyun.dev.run_remote_shell --script_path ${SCRIPT_FILE} --without_stop ${WITHOUT_STOP} --execute_user webuser --image_id "m-bp11q5jxgj4921h8ue1h"
+python -m pymlsql.aliyun.dev.run_remote_shell --script_path ${SCRIPT_FILE} --keyPairName "mlsql-build-env-local" --execute_user webuser --instance_id $1
